refactor(mobile): rename bundle id to org.visaliacrc.rhythm.mobile
Renames the iOS bundle identifier, Android applicationId, URL scheme,
and Kotlin package from com.rhythmapp.mobile to org.visaliacrc.rhythm.mobile
to align with the existing org.visaliacrc.rhythm namespace used by the
desktop app. Updates Info.plist URL scheme, project.pbxproj all three
build configs, build.gradle.kts namespace/applicationId/manifestPlaceholders,
and physically moves MainActivity.kt with history-preserving git mv.
