+++ apps/mobile_flutter/android/app/src/main/kotlin/org/visaliacrc/rhythm/mobile/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.rhythmapp.mobile.mobile_flutter
+package org.visaliacrc.rhythm.mobile
 
 import io.flutter.embedding.android.FlutterActivity
