+++ apps/mobile_flutter/ios/Runner/Info.plist
@@ -53,10 +53,10 @@
 			<key>CFBundleTypeRole</key>
 			<string>Editor</string>
 			<key>CFBundleURLName</key>
-			<string>com.rhythmapp.mobile</string>
+			<string>org.visaliacrc.rhythm.mobile</string>
 			<key>CFBundleURLSchemes</key>
 			<array>
-				<string>com.rhythmapp.mobile</string>
+				<string>org.visaliacrc.rhythm.mobile</string>
 			</array>
 		</dict>
 	</array>
